requirements.txt
absl-py==2.2.2
einops==0.8.1
email-validator==2.2.0
finnhub-python==2.4.24
h5py==3.13.0
httptools==0.6.4
keopscore==2.3
llama-cpp-python==0.3.8
matplotlib==3.9.4
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.4.0
patool==1.15.0
peft==0.17.1
pytorch-wavelet==1.2
PyWavelets==1.6.0
scikit-image==0.24.0
scikit-learn==1.6.1
seaborn==0.13.2
statsmodels==0.14.5
ta-lib==0.6.4
transformers==4.55.4
tokenizers==0.21.4
trl==0.22.2
accelerate==1.10.1
unsloth==2025.9.7
unsloth_zoo==2025.9.9
vllm==0.10.2
datasets==3.6.0
bitsandbytes==0.47.0
sentencepiece
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yfinance==0.2.56
wandb==0.22.0